+++ crates/primitives/src/basefee.rs
@@ -1,16 +1,15 @@
-use alloy_eips::eip1559::BaseFeeParams;
-use reth_primitives::basefee::calc_next_block_base_fee;
+use alloy_eips::eip1559::{calc_next_block_base_fee, BaseFeeParams};
 
 use crate::MIN_BASE_FEE_PER_GAS;
 
 pub fn calculate_next_block_base_fee(
-    gas_used: u128,
-    gas_limit: u128,
+    gas_used: u64,
+    gas_limit: u64,
     base_fee: Option<u64>,
     base_fee_params: BaseFeeParams,
 ) -> Option<u64> {
     Some(std::cmp::max(
         MIN_BASE_FEE_PER_GAS,
-        calc_next_block_base_fee(gas_used, gas_limit, base_fee? as u128, base_fee_params),
+        calc_next_block_base_fee(gas_used, gas_limit, base_fee?, base_fee_params),
     ) as u64)
 }
